requirements.txt
annotated-types==0.6.0
anyio==4.3.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.5.2
asttokens==2.0.5
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
binaryornot==0.4.4
bleach==5.0.1
blinker==1.7.0
blis==0.7.11
bson==0.5.10
cachetools==5.3.3
catalogue==2.0.10
certifi==2024.2.2
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
confection==0.1.4
control==0.9.4
cookiecutter==2.6.0
cycler==0.11.0
cymem==2.0.8
debugpy==1.6.2
decorator==4.4.2
defusedxml==0.7.1
distro==1.9.0
Django==4.1
dnspython==2.6.1
entrypoints==0.4
exceptiongroup==1.2.0
executing==0.9.1
fastapi==0.110.0
fastjsonschema==2.16.1
Flask==3.0.2
flet==0.21.1
flet-core==0.21.1
flet-runtime==0.21.1
fonttools==4.36.0
gcloud==0.18.3
google-ai-generativelanguage==0.4.0
google-api-core==2.17.1
google-auth==2.28.2
google-cloud-aiplatform==1.44.0
google-cloud-bigquery==3.19.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.3
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-generativeai==0.4.1
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
graphh==0.1.3
greenlet==2.0.2
grpc-google-iam-v1==0.13.0
grpcio==1.62.1
grpcio-status==1.62.1
h11==0.14.0
httpcore==1.0.4
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
idna==3.6
image==1.5.33
imageio==2.22.4
imageio-ffmpeg==0.4.7
ipykernel==6.15.1
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.1
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.9.1
jupyter==1.0.0
jupyter-client==7.3.4
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.1
kiwisolver==1.4.4
langcodes==3.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.5.3
matplotlib-inline==0.1.3
mdurl==0.1.2
mistune==0.8.4
moviepy==1.0.3
murmurhash==1.0.10
mysql==0.0.3
mysql-client==0.0.1
mysql-connector==2.2.9
mysql-connector-python==8.0.32
mysqlclient==2.1.1
nbclient==0.6.6
nbconvert==6.5.0
nbformat==5.4.0
neo4j==5.8.1
nest-asyncio==1.5.5
networkx==3.1
notebook==6.4.12
numpy==1.26.4
oauth2client==4.1.3
oauthlib==3.2.2
openai==1.13.3
packaging==23.2
paho-mqtt==1.6.1
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
path==16.6.0
pickleshare==0.7.5
Pillow==9.2.0
preshed==3.0.9
proglog==0.1.10
prometheus-client==0.14.1
prompt-toolkit==3.0.30
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.1
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
pydataset==0.2.0
pygame==2.1.2
Pygments==2.17.2
pymongo==4.6.2
PyMySQL==1.0.2
pyodbc==4.0.35
pyparsing==3.1.2
pypng==0.20220715.0
pyrsistent==0.18.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pytz==2022.7.1
pywin32==304
pywinpty==2.0.6
PyYAML==6.0.1
pyzmq==23.2.0
qrcode==7.4.2
qtconsole==5.3.1
QtPy==2.1.0
repath==0.9.0
requests==2.31.0
rich==13.7.1
rsa==4.9
scapy==2.5.0
scapy-python3==0.26
scikit-fuzzy==0.4.2
scikit-learn==1.4.1.post1
scipy==1.12.0
seaborn==0.12.2
Send2Trash==1.8.0
shapely==2.0.3
six==1.16.0
smart-open==6.4.0
sniffio==1.3.1
sounddevice==0.4.5
soupsieve==2.3.2.post1
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.7
sqlparse==0.4.2
srsly==2.4.8
stack-data==0.3.0
starlette==0.36.3
termcolor==0.2
terminado==0.15.0
text-unidecode==1.3
thinc==8.2.3
threadpoolctl==3.4.0
tinycss2==1.1.1
tornado==6.2
tqdm==4.66.2
traitlets==5.3.0
typer==0.9.0
types-python-dateutil==2.9.0.20240315
typing_extensions==4.10.0
tzdata==2022.2
urllib3==2.2.1
uvicorn==0.28.0
wasabi==1.1.2
watchdog==4.0.0
watchfiles==0.21.0
wcwidth==0.2.5
weasel==0.3.4
webencodings==0.5.1
websockets==12.0
Werkzeug==3.0.1
widgetsnbextension==3.6.1